requirements/base.txt
click==6.7
PyESCPOS==0.0.13
